DMOH-10: Added Channel HtmlHandler that handles all HtmlRequests.
