+++ tests/Handler/Channel/OpenNowHtmlHandlerTest.php
@@ -26,27 +26,4 @@ class OpenNowHtmlHandlerTest extends HandlerTestBase
             'Handler only handles \StadGent\Services\OpeningHours\Request\Channel\OpenNowRequest.'
         );
     }
-
-    /**
-     * Test the toResponse method with returned data.
-     */
-    public function testToResponseWithData()
-    {
-        $body = <<<EOT
-<div vocab="http://schema.org/" typeof="Library">
-    <h1>FooBar</h1>
-    <div>open</div>
-</div>
-EOT;
-        $openNowResponse = $this->createResponseMock(200, $body);
-
-        $handler = new OpenNowHtmlHandler();
-        $response = $handler->toResponse($openNowResponse);
-
-        $this->assertInstanceOf(
-            HtmlResponse::class,
-            $response,
-            'Response should be a \StadGent\Services\OpeningHours\Response\HtmlResponse object.'
-        );
-    }
 }
